Add functionality for updating `.yarnrc.yml` file (#2)
* Add functionality for updating the `.yarnrc.yml` file

The local `.yarnrc.yml` will now be updated to match the module template. This supports updating from Yarn 1 to Yarn 3, and updating to a more recent version of Yarn 3. It will also install any plugins specified in the module template.

* Remove comment in Caveats section

The `.yarnrc.yml` file is no longer skipped, as it will now be processed with some custom logic. The comment is now obsolete.

* Print warning and delete old `.yarnrc` if it exists

* Fix updating Yarn from 1 to 3

Previously, Yarn 1 would be set as `yarnPath` in the configuration file, resulting in errors when running Yarn.

* Simplify Yarn installation

* Add missing dependency

* Update documentation comment